requirements.txt
# DC Collectors Requirements
# Python 3.8+

# Web scraping
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0

# Data processing
pandas>=2.0.0
tqdm>=4.66.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
urllib3>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0